% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_files.R
\name{create_vignette_last_modif}
\alias{create_vignette_last_modif}
\title{Create the vignette for last modification}
\usage{
create_vignette_last_modif(
  repo = ".",
  path = "R",
  recursive = TRUE,
  untracked = TRUE
)
}
\arguments{
\item{repo}{git repository}

\item{path}{Default to R folder. Use "" for the complete repository.}

\item{recursive}{Logical. Should the listing recurse into directories?}

\item{untracked}{Logical. Should the untracked files be included?}
}
\value{
copy a vignette
}
\description{
Create the vignette for last modification
}
\examples{
repo <- fake_repo(as.package = TRUE)
create_vignette_last_modif(repo, path = "R")
}
