% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{each_commit}
\alias{each_commit}
\title{Write presentation for each commit}
\usage{
each_commit(data, write_in)
}
\arguments{
\item{data}{list with pattern and data elements}

\item{write_in}{function to write in good place}
}
\description{
Write presentation for each commit
}
