% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{each_pattern}
\alias{each_pattern}
\title{Create the text to add in a markdown file to present each pattern as a chapter of the book}
\usage{
each_pattern(nest_commits, pattern.type)
}
\arguments{
\item{nest_commits}{commits as nested with nest_commits_by_pattern}

\item{pattern.type}{Character name of the pattern to filter}
}
\value{
A tibble with a row for each different pattern found and
a 'text' column to be included in a markdown file.
}
\description{
Create the text to add in a markdown file to present each pattern as a chapter of the book
}
\examples{
repo <- fake_repo()
res_commits <- nest_commits_by_pattern(
  repo,
  pattern = c("Tickets" = "ticket[[:digit:]]+"),
  ref = "master", silent = TRUE
)
each_pattern(res_commits, "Tickets")
}
