% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_repo.R
\name{fake_repo}
\alias{fake_repo}
\title{Create a fake git repository in tempfile for examples}
\usage{
fake_repo(path = tempfile(pattern = "git2r-"))
}
\arguments{
\item{path}{Path to fake repository}
}
\value{
path to fake repository
}
\description{
Create a fake git repository in tempfile for examples
}
\examples{
fake_repo()
}
