% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_files.R
\name{get_info}
\alias{get_info}
\title{Get the first and last modification time for a specific file, based on git2r::blame().}
\usage{
get_info(path, repo = ".")
}
\arguments{
\item{path}{path to the file}

\item{repo}{repo of the git project}
}
\value{
A list with information on first and last modification time of the selected file.
}
\description{
Get the first and last modification time for a specific file, based on git2r::blame().
}
\examples{
repo <- fake_repo()
get_info(list.files(repo)[1], repo = repo)
}
