% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitdown.R
\name{git_copy_repo}
\alias{git_copy_repo}
\title{Copy a repository from GitHub}
\usage{
git_copy_repo(github_repo, copy_to = "repos")
}
\arguments{
\item{github_repo}{The user and repository name}

\item{copy_to}{Subfolder in the project where the repo will be copied to. Defaults to: "repos"}
}
\description{
Copy a repository from GitHub
}
\examples{
git_copy_repo("tidyverse/dbplyr")

git_copy_file("dbplyr/vignettes/dbplyr.Rmd",
             "content/dplyr.Rmd")

git_replace_text("content/dplyr.Rmd",
             "Introduction to dbplyr",
             "Databases using dplyr" )

git_copy_file("dbplyr/vignettes/sql-translation.Rmd",
             "content/translation.Rmd")

git_remove_lines("content/translation.Rmd", 247, 248)

}
