% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitdown.R
\name{git_remove_lines}
\alias{git_remove_lines}
\title{Removes a range numbered lines from a file}
\usage{
git_remove_lines(location, start, end)
}
\arguments{
\item{location}{Path to the file to be modified}

\item{start}{First line to be removed}

\item{end}{Last line to be removed}
}
\description{
Removes a range numbered lines from a file
}
\examples{
git_copy_repo("tidyverse/dbplyr")

git_copy_file("dbplyr/vignettes/dbplyr.Rmd",
             "content/dplyr.Rmd")

git_replace_text("content/dplyr.Rmd",
             "Introduction to dbplyr",
             "Databases using dplyr" )

git_copy_file("dbplyr/vignettes/sql-translation.Rmd",
             "content/translation.Rmd")

git_remove_lines("content/translation.Rmd", 247, 248)

}
