% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitdown-package.R
\docType{package}
\name{gitdown-package}
\alias{gitdown}
\alias{gitdown-package}
\title{gitdown: Turn Your Git Commit Messages into a HTML Book}
\description{
Read all commit messages of your local git repository
and sort them according to tags or specific text pattern into chapters
of a HTML book using 'bookdown'.
The git history book presentation helps organisms required to testify
for every changes in their source code, in relation to features requests.

The main functions to be used in this package are \code{\link[=git_down]{git_down()}} and \code{\link[=create_vignette_last_modif]{create_vignette_last_modif()}}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://thinkr-open.github.io/gitdown/}
  \item \url{https://github.com/Thinkr-open/gitdown}
  \item Report bugs at \url{https://github.com/Thinkr-open/gitdown/issues}
}

}
\author{
\strong{Maintainer}: Sébastien Rochette \email{sebastien@thinkr.fr} (\href{https://orcid.org/0000-0002-1565-9313}{ORCID})

Authors:
\itemize{
  \item Cervan Girard \email{cervan@thinkr.fr} (\href{https://orcid.org/0000-0002-4816-4624}{ORCID})
}

Other contributors:
\itemize{
  \item ThinkR [copyright holder]
  \item Institut de Recherches Internationales Servier [sponsor]
}

}
\keyword{internal}
