% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_files.R
\name{mise_en_forme}
\alias{mise_en_forme}
\title{Formatting results of get_last_modif}
\usage{
mise_en_forme(repo = ".", R_folder = TRUE)
}
\arguments{
\item{repo}{repo}

\item{R_folder}{If TRUE, it will list functions inside the R folder}
}
\value{
tagList
}
\description{
Formatting results of get_last_modif
}
\examples{
repo <- fake_repo()
mise_en_forme(repo, R_folder = FALSE)
}
