% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_commits.R
\name{my_extract}
\alias{my_extract}
\title{Extract pattern from message}
\usage{
my_extract(message, pattern)
}
\arguments{
\item{message}{Character}

\item{pattern}{character vector;
search patterns; for more details refer to \link[stringi]{stringi-search}}
}
\description{
Extract pattern from message
}
