% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nest_commits_by_pattern}
\alias{nest_commits_by_pattern}
\title{Nest all commits by each pattern found in the commit messages}
\usage{
nest_commits_by_pattern(
  repo,
  pattern = c(Issues = "#[[:digit:]]+"),
  pattern.table = NULL,
  ref = "master",
  silent = TRUE
)
}
\arguments{
\item{repo}{a path to a repository or a \code{git_repository}
object. Default is '.'}

\item{pattern}{Named vector with regex pattern to expose commits, like c("Issues" = "#[[:digit:]]") for issues}

\item{pattern.table}{data.frame with two columns: pattern and description.
This is used as correspondence table to add some names to existing patterns.}

\item{ref}{The name of a reference to list commits from e.g. a tag
or a branch. The default is NULL for the current branch.}

\item{silent}{Logical. Whether to hide messages.}
}
\value{
A tibble with a row for each different pattern found and
a nested 'data' column with all related commits
}
\description{
Nest all commits by each pattern found in the commit messages
}
\examples{
repo <- fake_repo()
nest_commits_by_pattern(repo)
}
