% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_files.R
\name{present_files}
\alias{present_files}
\title{Formatting results of get_last_modif}
\usage{
present_files(repo = ".", path = "R", recursive = TRUE, untracked = TRUE)
}
\arguments{
\item{repo}{git repository}

\item{path}{Default to R folder. Use "" for the complete repository.}

\item{recursive}{Logical. Should the listing recurse into directories?}

\item{untracked}{Logical. Should the untracked files be included?}
}
\value{
kable
}
\description{
Formatting results of get_last_modif
}
\examples{
repo <- fake_repo(as.package = TRUE)
cat(present_files(repo))
}
