% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{presentation_commit}
\alias{presentation_commit}
\title{Presentation of commit}
\usage{
presentation_commit(commit)
}
\arguments{
\item{commit}{one line data from get_commits_pattern}
}
\description{
Presentation of commit
}
