% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_in_file}
\alias{replace_in_file}
\title{Replace in file}
\usage{
replace_in_file(file, pattern, replacement)
}
\arguments{
\item{file}{Path of file}

\item{pattern}{Pattern to replace}

\item{replacement}{A character vector of replacements}
}
\description{
Replace in file
}
