% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_files.R
\name{update_vign_last_modif}
\alias{update_vign_last_modif}
\title{Update vignette last modification of files}
\usage{
update_vign_last_modif(repo = ".", R_folder = TRUE)
}
\arguments{
\item{repo}{repo}

\item{R_folder}{If TRUE, it will list functions inside the R folder}
}
\value{
update a vignette
}
\description{
Update vignette last modification of files
}
