% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_files.R
\name{update_vignette_last_modif}
\alias{update_vignette_last_modif}
\title{Update vignette last modification of files}
\usage{
update_vignette_last_modif(
  repo = ".",
  path = "R",
  recursive = TRUE,
  untracked = TRUE
)
}
\arguments{
\item{repo}{git repository}

\item{path}{Default to R folder. Use "" for the complete repository.}

\item{recursive}{Logical. Should the listing recurse into directories?}

\item{untracked}{Logical. Should the untracked files be included?}
}
\value{
update existing vignette
}
\description{
Update vignette last modification of files
}
\examples{
repo <- fake_repo(as.package = TRUE)
update_vignette_last_modif(repo, path = "R")
rmarkdown::render(file.path(repo, "vignettes", "modification_files.Rmd"))
}
