% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_files.R
\name{vignette_last_modif}
\alias{vignette_last_modif}
\title{Get the vignette for last modification}
\usage{
vignette_last_modif(repo = ".")
}
\arguments{
\item{repo}{repo}
}
\value{
copy a vignette
}
\description{
Get the vignette for last modification
}
