% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_in}
\alias{write_in}
\title{Write inside file}
\usage{
write_in(x, repo, dir = "gitdown", rmd = "index.Rmd")
}
\arguments{
\item{x}{what}

\item{repo}{where}

\item{dir}{by deflaut gitbook}

\item{rmd}{wich file}
}
\description{
Write inside file
}
