% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{each_pattern}
\alias{each_pattern}
\title{Presentation for each pattern}
\usage{
each_pattern(pattern, name_section, repo, ref = "master")
}
\arguments{
\item{pattern}{vector of regex patterns}

\item{name_section}{names for pattern section}

\item{repo}{path to the repo}

\item{ref}{branch of the repo}
}
\description{
Presentation for each pattern
}
\examples{
repo <- fake_repo()
each_pattern("ticket[[:digit:]]+", "Ticket", repo)
}
