% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{each_pattern}
\alias{each_pattern}
\title{Presentation for each pattern}
\usage{
each_pattern(nest_commits, pattern.type)
}
\arguments{
\item{nest_commits}{commits as nested with nest_commits_by_pattern}

\item{pattern.type}{Character name of the pattern to filter}
}
\description{
Presentation for each pattern
}
\examples{
repo <- fake_repo()
res_commits <- nest_commits_by_pattern(
  repo,
  pattern = c("Tickets" = "ticket[[:digit:]]+"),
  ref = "master", silent = TRUE
)
each_pattern(res_commits, "Tickets")
}
