% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{each_pattern}
\alias{each_pattern}
\title{Presentation for each pattern}
\usage{
each_pattern(vec_pattern, names_section, repo, ref = "master", write_in)
}
\arguments{
\item{vec_pattern}{vector of pattern}

\item{names_section}{names for pattern section}

\item{repo}{path to the repo}

\item{ref}{branch of the repo}

\item{write_in}{function to write in good place}
}
\description{
Presentation for each pattern
}
