% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_repo.R
\name{fake_repo}
\alias{fake_repo}
\title{Create a fake git repository in a specific folder}
\usage{
fake_repo(path = tempfile(pattern = "git2r-"), as.package = FALSE)
}
\arguments{
\item{path}{Path to fake repository}

\item{as.package}{Logical. Whether to add R/ and vignettes/
directories to fake a package}
}
\value{
Character. Path of a fake repository used for reproducible examples.
Fake repository contains a few files with an initiated git repository.
}
\description{
A fake repository allows to create reproducible examples for this package functions
}
\examples{
# Fake repository with git
fake_repo()
# Fake repository that looks like package with git
fake_repo(as.package = TRUE)
}
