% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_files.R
\name{get_info}
\alias{get_info}
\title{Get info}
\usage{
get_info(path, repo = ".")
}
\arguments{
\item{path}{path to the file}

\item{repo}{repo of the git project}
}
\value{
list
}
\description{
Get the first and last modification for one file. Based on git2r::blame.
}
\examples{
repo <- fake_repo()
get_info(list.files(repo)[1], repo = repo )
}
