% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_files.R
\name{get_last_modif}
\alias{get_last_modif}
\title{Get last modification of files}
\usage{
get_last_modif(repo = ".", R_folder = TRUE)
}
\arguments{
\item{repo}{repo}

\item{R_folder}{If TRUE, it will list functions inside the R folder}
}
\value{
list
}
\description{
Get last modification of files
}
\examples{
repo <- fake_repo()
get_last_modif(repo = repo , R_folder = FALSE)
}
