% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_files.R
\name{get_last_modif}
\alias{get_last_modif}
\title{Get the first and last modification time of files of a repository}
\usage{
get_last_modif(repo = ".", path = "R", recursive = TRUE, untracked = TRUE)
}
\arguments{
\item{repo}{git repository}

\item{path}{Default to R folder. Use "" for the complete repository.}

\item{recursive}{Logical. Should the listing recurse into directories?}

\item{untracked}{Logical. Should the not tracked files be included?}
}
\value{
A list of files with information on first and last modification time.
}
\description{
Get the first and last modification time of files of a repository
}
\examples{
repo <- fake_repo()
# Complete repository
get_last_modif(repo = repo, path = "")
repo <- fake_repo(as.package = TRUE)
# Complete repository
get_last_modif(repo = repo, path = "")
}
