% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_down.R
\name{git_down}
\alias{git_down}
\title{git history to bookdown}
\usage{
git_down(
  repo = ".",
  book_path = "gitdown",
  open = TRUE,
  author = "John Doe",
  pattern = c(Issues = "#[[:digit:]]+"),
  ref = "master",
  ...
)
}
\arguments{
\item{repo}{The path to a repository. Default is `.`}

\item{book_path}{The path to the bookdown output. Default is `"gitdown"`.}

\item{open}{Should the bookdown be opened once compiled? Default is TRUE.}

\item{author}{Author of the Bookdown}

\item{pattern}{Named vector with regex pattern to expose commits, like c("Issues" = "#[[:digit:]]") for issues}

\item{ref}{the name of the branch, by default master}

\item{...}{Other parameters to pass to \code{\link[rmarkdown]{render}}}
}
\description{
This function turns the history of git into a bookdown.
Each chapter is a group of commits. The first page gives a summary of all the groups.
}
\examples{
repo <- fake_repo()
git_down(repo, pattern = c("Tickets" = "ticket[[:digit:]]+", "Issues" = "#[[:digit:]]+"))
}
