% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_down.R
\name{git_down}
\alias{git_down}
\title{git history to bookdown}
\usage{
git_down(repo = ".", book_path = "gitdown", open = TRUE)
}
\arguments{
\item{repo}{The path to a repository. Default is `.`}

\item{book_path}{The path to the bookdown output. Default is `"gitdown"`.}

\item{open}{Should the bookdown be opened once compiled? Default is TRUE.}
}
\description{
This function turns the history of git into a bookdown.
Each chapter is a group of commits. The first page gives a summary of all the groups.
}
