% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{nest_commits_by_pattern}
\alias{nest_commits_by_pattern}
\title{Nest all commits by each pattern}
\usage{
nest_commits_by_pattern(repo, pattern, ref, silent = TRUE)
}
\arguments{
\item{repo}{a path to a repository or a \code{git_repository}
object. Default is '.'}

\item{pattern}{Pattern to look for.

The default interpretation is a regular expression, as described
in \link[stringi:stringi-search-regex]{stringi::stringi-search-regex}. Control options with
\code{\link[stringr:regex]{regex()}}.

Match a fixed string (i.e. by comparing only bytes), using
\code{\link[stringr:fixed]{fixed()}}. This is fast, but approximate. Generally,
for matching human text, you'll want \code{\link[stringr:coll]{coll()}} which
respects character matching rules for the specified locale.

Match character, word, line and sentence boundaries with
\code{\link[stringr:boundary]{boundary()}}. An empty pattern, "", is equivalent to
\code{boundary("character")}.}

\item{ref}{The name of a reference to list commits from e.g. a tag
or a branch. The default is NULL for the current branch.}

\item{silent}{Logical. Whether to hide messages.}
}
\description{
Nest all commits by each pattern
}
