% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_info_files.R
\name{present_files}
\alias{present_files}
\title{Presenting results of files and last modification time in a printed table using 'kable()'}
\usage{
present_files(repo = ".", path = "R", recursive = TRUE, untracked = TRUE)
}
\arguments{
\item{repo}{git repository}

\item{path}{Default to R folder. Use "" for the complete directory}

\item{recursive}{Logical. Should the listing recurse into directories?}

\item{untracked}{Logical. Should the not tracked files be included?}
}
\value{
A 'kable()' output to be included in a markdown file
}
\description{
Presenting results of files and last modification time in a printed table using 'kable()'
}
\examples{
repo <- fake_repo(as.package = TRUE)
cat(present_files(repo))
}
