% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{presentation_commit}
\alias{presentation_commit}
\title{Presentation of commit}
\usage{
presentation_commit(write_in, data, i)
}
\arguments{
\item{write_in}{function to write in good place}

\item{data}{data from get_commits_pattern}

\item{i}{line number to present the commit}
}
\description{
Presentation of commit
}
