% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{write_in}
\alias{write_in}
\title{Write inside file}
\usage{
write_in(x, repo, dir = "gitdown", rmd = "index.Rmd")
}
\arguments{
\item{x}{what text to write}

\item{repo}{Path to directory where to store "dir"}

\item{dir}{Name of the dir where to write file}

\item{rmd}{name and type of the file to write in}
}
\description{
Write inside file
}
